Implemented new auth flow

+++ src/python/api/apis/endpoints/ngroup.py
@@ -1,30 +1,41 @@
 from fastapi import APIRouter, HTTPException, Query, Depends, Request
 from uuid import UUID
-from typing import List, Optional
+from typing import List, Optional, Dict, Any
 
 from utils.ngroup import create_ngroup, get_ngroup_id_by_name, get_ngroup, update_ngroup, delete_ngroup, list_ngroups, NgroupNotFoundError  
 from lambda_utils.type_util.ngroup import NgroupCreate, NgroupReturn, NgroupUpdate
 
-from lambda_utils.type_util.cueuser import CueuserReturn
+from utils.JWTBearer import bearer_scheme, JWTBearer  # Import the dependency
+from utils.cognito_utils import cognito_auth
 
-from utils.auth import get_current_user
 
 router = APIRouter(prefix="/ngroup", tags=["ngroup"])
 
 @router.post("/", response_model=NgroupReturn)
-async def create_ngroup_endpoint(ngroup: NgroupCreate,  request: Request,  user: CueuserReturn = Depends(get_current_user)):
-    print("Query Params:", request.query_params)
-    print("Path Params:", request.path_params)
-    body = await request.body()
-    print("Body : ", body)
-    return await create_ngroup(ngroup)
+async def create_ngroup_endpoint(
+    ngroup: NgroupCreate,
+    current_user: dict = Depends(cognito_auth.get_current_user),
+    token: str = Depends(bearer_scheme)
+):
+    if not current_user:
+        raise HTTPException(status_code=401, detail="Not authenticated")
+    try:
+        return await create_ngroup(ngroup)
+    except ValueError as e:
+        raise HTTPException(status_code=400, detail=str(e))
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"An unexpected error occurred: {str(e)}")
 
 @router.get("/id", response_model=UUID)
 async def get_ngroup_id_endpoint(
     short_name: Optional[str] = Query(None, alias="short_name"),
     long_name: Optional[str] = Query(None, alias="long_name"),
-    user: CueuserReturn = Depends(get_current_user)
+    current_user: dict = Depends(cognito_auth.get_current_user),
+    token: str = Depends(bearer_scheme)
+
 ):
+    if not current_user:
+        raise HTTPException(status_code=401, detail="Not authenticated")
     try:
         ngroup_id = await get_ngroup_id_by_name(short_name, long_name)
         return ngroup_id
@@ -34,7 +45,13 @@ async def get_ngroup_id_endpoint(
         raise HTTPException(status_code=404, detail=str(e))
 
 @router.get("/{ngroup_id}", response_model=NgroupReturn)
-async def get_ngroup_endpoint(ngroup_id: UUID, user: CueuserReturn = Depends(get_current_user)):
+async def get_ngroup_endpoint(
+    ngroup_id: UUID,
+    current_user: dict = Depends(cognito_auth.get_current_user),
+    token: str = Depends(bearer_scheme)
+):
+    if not current_user:
+        raise HTTPException(status_code=401, detail="Not authenticated")
     try:
         ngroup = await get_ngroup(ngroup_id)
         return ngroup
@@ -42,7 +59,14 @@ async def get_ngroup_endpoint(ngroup_id: UUID, user: CueuserReturn = Depends(get
         raise HTTPException(status_code=404, detail=str(e))
 
 @router.patch("/{ngroup_id}", response_model=NgroupReturn)
-async def update_ngroup_endpoint(ngroup_id: UUID, ngroup_update: NgroupUpdate, user: CueuserReturn = Depends(get_current_user)):
+async def update_ngroup_endpoint(
+    ngroup_id: UUID,
+    ngroup_update: NgroupUpdate,
+    current_user: dict = Depends(cognito_auth.get_current_user),
+    token: str = Depends(bearer_scheme)
+):
+    if not current_user:
+        raise HTTPException(status_code=401, detail="Not authenticated")
     try:
         updated_ngroup = await update_ngroup(ngroup_id, ngroup_update)
         return updated_ngroup
@@ -50,7 +74,13 @@ async def update_ngroup_endpoint(ngroup_id: UUID, ngroup_update: NgroupUpdate, u
         raise HTTPException(status_code=404, detail=str(e))
 
 @router.delete("/{ngroup_id}", response_model=bool)
-async def delete_ngroup_endpoint(ngroup_id: UUID, user: CueuserReturn = Depends(get_current_user)):
+async def delete_ngroup_endpoint(
+    ngroup_id: UUID,
+    current_user: dict = Depends(cognito_auth.get_current_user),
+    token: str = Depends(bearer_scheme)
+):
+    if not current_user:
+        raise HTTPException(status_code=401, detail="Not authenticated")
     try:
         success = await delete_ngroup(ngroup_id)
         return success
@@ -58,6 +88,8 @@ async def delete_ngroup_endpoint(ngroup_id: UUID, user: CueuserReturn = Depends(
         raise HTTPException(status_code=404, detail=str(e))
 
 @router.get("/", response_model=List[NgroupReturn])
-async def list_ngroups_endpoint(user: CueuserReturn = Depends(get_current_user)):
+async def list_ngroups_endpoint(current_user: dict = Depends(cognito_auth.get_current_user),
+    token: str = Depends(bearer_scheme)):
+    if not current_user:
+        raise HTTPException(status_code=401, detail="Not authenticated")
     return await list_ngroups()
-
